Simple swap to using Guzzle 3
This doesn't solve the issue in constructing Messagebird\Client as it just changes HttpClient to wrap Guzzle.
This solves an issue where headers are set for the Voice API client, which are not used because they're the wrong format (they're passed straight to cURL with no validation or filtering, and cURL expects an indexed array, not associative).
Overall this could still be vastly improved by dropping support for older PHP versions, and updating to Guzzle 6.
Could then add better unit tests, and possibly scrap HttpClient altogether if able to change the contracted API for HttpClient (which isn't documented anyway and isn't the primary interface for this package).
`addUserAgentString()`, `setHeaders()`, `setAuthentication()` are all methods that could have their implementation changed to operate directly on a Guzzle client.